Make build more modular
