+++ build.js
@@ -1,37 +1,42 @@
 const STATIC_FILES = ['favicon.ico', 'robots.txt', 'sitemap.xml', '_redirects'];
-const exstatic = require('./@exstatic/packages/dev');
-const {performance} = require('perf_hooks');
+const argString = process.argv.join('  ');
+const hasArg = arg => Boolean(new RegExp(`-?-${arg}`).test(argString));
 
-const instance = exstatic();
-
-function copyFiles() {
-	const {copy} = require('fs-extra');
+async function buildAssets() {
 	const {css} = require('./scripts/css');
 	const {js} = require('./scripts/js')
+	console.log('Building assets');
+	return Promise.all([css(), js()]);
+}
+
+async function copyFiles() {
+	const {copy} = require('fs-extra');
+
+	if (hasArg('build')) {
+		await buildAssets();
+	}
 
 	// Exstatic doesn't currently have support for 1:1 copying
-	const promises = [copy('./src/assets', './built/assets'), css(), js()];
+	const promises = [copy('./src/assets', './built/assets')];
 	STATIC_FILES.forEach(file => promises.push(copy(`./src/${file}`, `./built/${file}`)));
+	console.log('Copying assets');
 	return Promise.all(promises);
 }
 
-async function compile() {
-	const start = performance.now();
-	await instance.build();
-	const time = performance.now() - start;
-	console.log(`Build took ${(time / 1000).toFixed(3)}s`);
-	instance.onBeforeExit(true);
+function compile() {
+	const runCompile = require('./scripts/exstatic');
+	return runCompile();
 }
 
 async function run() {
-	console.log('Copying assets');
 	await copyFiles();
-	console.log('Building exstatic');
 	return compile();
 }
 
-if (process.argv.join(' ').match(/-?-exstatic/)) {
+if (hasArg('compile')) {
 	compile();
+} else if (hasArg('copy') || hasArg('build')){
+	copyFiles();
 } else {
 	run();
 }
